requirements.txt
# Python dependencies
datasets
google-cloud-storage
pandas
psycopg2-binary
sqlalchemy
google-cloud-aiplatform

